Mattermost Integration: add LateInit, robust IsUpToDate, safe webhook behavior, full test suite, controller fixes

+++ pkg/namespaced/controller/projects/integrationmattermost/controller.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-	http://www.apache.org/licenses/LICENSE-2.0
+    http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -28,6 +28,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/v2/pkg/reconciler/managed"
 	"github.com/crossplane/crossplane-runtime/v2/pkg/resource"
 	"github.com/crossplane/crossplane-runtime/v2/pkg/statemetrics"
+	"github.com/google/go-cmp/cmp"
 	gitlab "gitlab.com/gitlab-org/api/client-go"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -40,6 +41,7 @@ import (
 
 const (
 	errNotIntegrationMattermost = "managed resource is not a Gitlab integration mattermost custom resource"
+	errProjectIDMissing         = "ProjectID is missing"
 	errGetFailed                = "cannot get Gitlab integration mattermost"
 	errCreateFailed             = "cannot create Gitlab integration mattermost"
 	errUpdateFailed             = "cannot update Gitlab integration mattermost"
@@ -62,12 +64,19 @@ func SetupIntegrationMattermost(mgr ctrl.Manager, o controller.Options) error {
 		reconcilerOpts = append(reconcilerOpts, managed.WithManagementPolicies())
 	}
 
-	r := managed.NewReconciler(mgr,
+	r := managed.NewReconciler(
+		mgr,
 		resource.ManagedKind(v1alpha1.IntegrationMattermostGroupVersionKind),
-		reconcilerOpts...)
+		reconcilerOpts...,
+	)
 
 	if err := mgr.Add(statemetrics.NewMRStateRecorder(
-		mgr.GetClient(), o.Logger, o.MetricOptions.MRStateMetrics, &v1alpha1.IntegrationMattermostList{}, o.MetricOptions.PollStateMetricInterval)); err != nil {
+		mgr.GetClient(),
+		o.Logger,
+		o.MetricOptions.MRStateMetrics,
+		&v1alpha1.IntegrationMattermostList{},
+		o.MetricOptions.PollStateMetricInterval,
+	)); err != nil {
 		return err
 	}
 
@@ -87,14 +96,13 @@ func SetupIntegrationMattermostGated(mgr ctrl.Manager, o controller.Options) err
 	return nil
 }
 
-// connector is responsible for producing an ExternalClient for Gitlab Integration Mattermost.
+// connector produces an ExternalClient for GitLab Integration Mattermost.
 type connector struct {
 	kube              client.Client
 	newGitlabClientFn func(cfg common.Config) projects.MattermostClient
 }
 
-// Connect creates a new Gitlab client for the given managed resource.
-// It fetches the provider configuration and uses it to instantiate the client.
+// Connect creates a new GitLab client for the given managed resource.
 func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.ExternalClient, error) {
 	cr, ok := mg.(*v1alpha1.IntegrationMattermost)
 	if !ok {
@@ -107,13 +115,13 @@ func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.E
 	return &external{kube: c.kube, client: c.newGitlabClientFn(*cfg)}, nil
 }
 
-// external represents the external client for Gitlab Integration Mattermost.
+// external represents the external client for GitLab Integration Mattermost.
 type external struct {
 	kube   client.Client
 	client projects.MattermostClient
 }
 
-// applies desired configuration of Mattermost to the project
+// applyMattermost applies desired Mattermost settings to the GitLab project.
 func (e *external) applyMattermost(ctx context.Context, cr *v1alpha1.IntegrationMattermost) error {
 	_, _, err := e.client.SetMattermostService(
 		*cr.Spec.ForProvider.ProjectID,
@@ -123,17 +131,20 @@ func (e *external) applyMattermost(ctx context.Context, cr *v1alpha1.Integration
 	return err
 }
 
-// Observe checks if the Gitlab Integration Mattermost external resource exists and
-// if it is up to date.
+// Observe checks whether the external resource exists and whether it is up-to-date.
 func (e *external) Observe(ctx context.Context, mg resource.Managed) (managed.ExternalObservation, error) {
 	cr, ok := mg.(*v1alpha1.IntegrationMattermost)
 	if !ok {
 		return managed.ExternalObservation{}, errors.New(errNotIntegrationMattermost)
 	}
 
-	// If the resource is being deleted, return early to avoid updating status
+	// If the resource is being deleted, avoid updating status.
 	if meta.WasDeleted(cr) {
 		return managed.ExternalObservation{ResourceExists: false}, nil
+	}
+
+	if cr.Spec.ForProvider.ProjectID == nil {
+		return managed.ExternalObservation{}, errors.New(errProjectIDMissing)
 	}
 
 	mattermost, res, err := e.client.GetMattermostService(
@@ -146,25 +157,37 @@ func (e *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 		}
 		return managed.ExternalObservation{}, errors.Wrap(err, errGetFailed)
 	}
+	if mattermost == nil || mattermost.Properties == nil {
+		return managed.ExternalObservation{ResourceExists: false}, nil
+	}
 
+	// Late initialize spec from remote (mutates in place).
+	current := cr.Spec.ForProvider.DeepCopy()
+	projects.LateInitializeIntegrationMattermost(&cr.Spec.ForProvider, mattermost)
+
+	// Update status from the remote state.
 	cr.Status.AtProvider = projects.GenerateIntegrationMattermostObservation(mattermost)
 	cr.Status.SetConditions(xpv1.Available())
 
 	return managed.ExternalObservation{
 		ResourceExists:          true,
 		ResourceUpToDate:        projects.IsIntegrationMattermostUpToDate(&cr.Spec.ForProvider, mattermost),
-		ResourceLateInitialized: false,
+		ResourceLateInitialized: !cmp.Equal(current, &cr.Spec.ForProvider),
 	}, nil
 }
 
-// Create creates the external resource for Gitlab Integration Mattermost.
-// Upon creation, Gitlab Integration Mattermost cannot be uniquely identified,
-// so we simply call SetMattermostService with the desired parameters.
+// Create creates the external resource for GitLab Integration Mattermost.
+// The integration is configured by sending desired options directly.
 func (e *external) Create(ctx context.Context, mg resource.Managed) (managed.ExternalCreation, error) {
 	cr, ok := mg.(*v1alpha1.IntegrationMattermost)
 	if !ok {
 		return managed.ExternalCreation{}, errors.New(errNotIntegrationMattermost)
 	}
+
+	if cr.Spec.ForProvider.ProjectID == nil {
+		return managed.ExternalCreation{}, errors.New(errProjectIDMissing)
+	}
+
 	cr.Status.SetConditions(xpv1.Creating())
 
 	if err := e.applyMattermost(ctx, cr); err != nil {
@@ -174,41 +197,48 @@ func (e *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 }
 
 // Update updates the external resource to match the desired state.
-// As Gitlab Integration Mattermost cannot be uniquely identified, we simply call
-// SetMattermostService with the desired parameters.
 func (e *external) Update(ctx context.Context, mg resource.Managed) (managed.ExternalUpdate, error) {
 	cr, ok := mg.(*v1alpha1.IntegrationMattermost)
 	if !ok {
 		return managed.ExternalUpdate{}, errors.New(errNotIntegrationMattermost)
 	}
 
+	if cr.Spec.ForProvider.ProjectID == nil {
+		return managed.ExternalUpdate{}, errors.New(errProjectIDMissing)
+	}
+
+	// Set Creating condition to align with controller tests and convention.
 	cr.Status.SetConditions(xpv1.Creating())
 
 	if err := e.applyMattermost(ctx, cr); err != nil {
 		return managed.ExternalUpdate{}, errors.Wrap(err, errUpdateFailed)
 	}
-
 	return managed.ExternalUpdate{}, nil
 }
 
-// Delete removes the integration mattermost resource.
+// Delete removes the GitLab Mattermost integration.
 func (e *external) Delete(ctx context.Context, mg resource.Managed) (managed.ExternalDelete, error) {
 	cr, ok := mg.(*v1alpha1.IntegrationMattermost)
 	if !ok {
 		return managed.ExternalDelete{}, errors.New(errNotIntegrationMattermost)
 	}
 
+	if cr.Spec.ForProvider.ProjectID == nil {
+		return managed.ExternalDelete{}, errors.New(errProjectIDMissing)
+	}
+
+	// Do not set Deleting condition to match existing controller test expectations.
 	_, err := e.client.DeleteMattermostService(
 		*cr.Spec.ForProvider.ProjectID,
-		gitlab.WithContext(ctx))
+		gitlab.WithContext(ctx),
+	)
 	if err != nil {
 		return managed.ExternalDelete{}, errors.Wrap(err, errDeleteFailed)
 	}
-
 	return managed.ExternalDelete{}, nil
 }
 
+// Disconnect is a no-op required by the SDK interface.
 func (e *external) Disconnect(ctx context.Context) error {
-	// Disconnect is not implemented as it is a new method required by the SDK
 	return nil
 }
